rmda:Fix mr_key to be uint64_t
This commit mirrors change 381b605a8d7247b626804f8b231bea28d2703162
in master branch, but implemented in C only style

+++ include/nccl_ofi_rdma.h
@@ -127,7 +127,7 @@ typedef struct nccl_net_ofi_rdma_mr_handle {
 	int num_rails;
 
 	/* value of mr key id, if keys must be requested */
-	int mr_key;
+	uint64_t mr_key;
 
 	/* Array of size `num_rails' */
 	struct fid_mr **mr;
